% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lukedatacleaning.R
\name{clean_data}
\alias{clean_data}
\title{Import and clean data from twitter relating to Donald Trump}
\usage{
clean_data()
}
\value{
Tidy data of tweets broken down by word and tweet index
}
\description{
Import and clean data from twitter relating to Donald Trump
}
