% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacycloud.R
\name{get_cloud}
\alias{get_cloud}
\title{Get wordcoud}
\usage{
get_cloud(num_words, final_scoring)
}
\arguments{
\item{num_words}{number of words desired in the cloud}

\item{final_scoring}{}
}
\value{
visualization of word cloud
}
\description{
Get wordcoud
}
