% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_locations.R
\name{lat_lon}
\alias{lat_lon}
\title{Retrieve location data for tweets about Donald Trump}
\usage{
lat_lon()
}
\value{
Data of latitude and longitudes of tweet data
}
\description{
Retrieve location data for tweets about Donald Trump
}
