% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lukesentiments.R
\name{sentiments_score}
\alias{sentiments_score}
\title{Assign average sentiment scores to each word in a tweet based on the average score for the tweet as a whole}
\usage{
sentiments_score(word_bag)
}
\arguments{
\item{word_bag}{data frame of individual words (indexed by tweet)}
}
\value{
Tidy data of tweets broken down by word and tweet index
}
\description{
Assign average sentiment scores to each word in a tweet based on the average score for the tweet as a whole
}
