% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_retrieval.R
\name{tweets}
\alias{tweets}
\title{Import text relating to Donald Trump}
\usage{
tweets()
}
\value{
Data frame of relevant information regarding tweets about Donald Trump
}
\description{
Import text relating to Donald Trump
}
